rename Batch.svelte to BatchChildBountyCalls
